[Generator] Move the conversion attr out and enable nullability.
Cleaned some code that could be simpler too.
